% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_generate_data.R
\name{generate_bivariate_normal_data_count}
\alias{generate_bivariate_normal_data_count}
\title{Generate data from bivariate normal covariates (count outcome)}
\usage{
generate_bivariate_normal_data_count(n, rho, alphas, gamma, sigma, beta0, betaA,
  betaX, betaXA)
}
\arguments{
\item{n}{Sample size}

\item{rho}{correlation coefficient}

\item{alphas}{True coefficients for the first and second treatment linear predictors. This vector should contain the intercept. alphas = c(alpha01, alphaXm1, alpha02, alphaXm2)}

\item{gamma}{Ratio of the true coefficient for the second variable and the first variable.}

\item{sigma}{scaling of all covariate effects. A higher value means stronger covariate effects, i.e., less clinical equipoise.}

\item{beta0}{Outcome model intercept coefficient}

\item{betaA}{Outcome model coefficient for I(A_i = 1) and I(A_i = 2)}

\item{betaX}{Outcome model coefficient vector for covariates X_i}

\item{betaXA}{Outcome model interaction coefficients for covariates. betaXA = c(betaXA1, betaXA2)}
}
\value{
a complete simulated data_frame
}
\description{
.. content for details ..
}
\author{
Kazuki Yoshida
}
