% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_generate_data.R
\name{generate_bivariate_standard_normal_covariate}
\alias{generate_bivariate_standard_normal_covariate}
\title{Generate bivariate standard normal distribution with correlation rho}
\usage{
generate_bivariate_standard_normal_covariate(n, rho)
}
\arguments{
\item{n}{Sample size}

\item{rho}{Correlation coefficient between X1 and X2}
}
\value{
data_frame containing two covariates X1 and X2. Both are marginally N(0,1). Their correlation is rho.
}
\description{
.. content for details ..
}
