% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_generate_data.R
\name{generate_p_dimensional_cont_count_bin_covariates}
\alias{generate_p_dimensional_cont_count_bin_covariates}
\title{Generate p covariate, one continuous, one count, and binary}
\usage{
generate_p_dimensional_cont_count_bin_covariates(n, p, rho, lambda, prev)
}
\arguments{
\item{n}{Sample size}

\item{p}{Dimension}

\item{rho}{Correlation coefficient. corr(Xi, Xj) = rho^abs(i - j) for the latent p-variate normal distribution.}

\item{lambda}{mean parameter for Poisson X2 variable}

\item{prev}{prevalence parameter vector for remaining binary variables. This must be p - 2.}
}
\value{
data_frame containing p covariates X1 through Xp. Each one is marginally N(0,1). Their correlation structure is compound symmetry.
}
\description{
.. content for details ..
}
