% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_generate_data.R
\name{generate_p_dimensional_standard_normal_covariates}
\alias{generate_p_dimensional_standard_normal_covariates}
\title{Generate p-variate standard normal distribution with correlation rho}
\usage{
generate_p_dimensional_standard_normal_covariates(n, p, rho)
}
\arguments{
\item{n}{Sample size}

\item{p}{Dimension}

\item{rho}{Correlation coefficient. corr(Xi, Xj) = rho^abs(i - j).}
}
\value{
data_frame containing p covariates X1 through Xp. Each one is marginally N(0,1). Their correlation structure is compound symmetry.
}
\description{
.. content for details ..
}
