% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_generate_data.R
\name{generate_p_norm_count_bin_data_count}
\alias{generate_p_norm_count_bin_data_count}
\title{Generate data from p-variate normal covariates (count outcome)}
\usage{
generate_p_norm_count_bin_data_count(n, p, rho, lambda, prev, alphas, sigma,
  beta0, betaA, betaX, betaXA)
}
\arguments{
\item{n}{Sample size}

\item{p}{number of covariates. Must be 3 or more.}

\item{rho}{correlation coefficient}

\item{lambda}{mean parameter for X2 (count variable)}

\item{prev}{prevalence vector for X3 through Xp (binary variables)}

\item{alphas}{True coefficients for the first and second treatment linear predictors. This vector should contain the intercept. alphas = c(alpha01, alphaXm1, alpha02, alphaXm2)}

\item{sigma}{scaling of all covariate effects. A higher value means stronger covariate effects, i.e., less clinical equipoise.}

\item{beta0}{Outcome model intercept coefficient}

\item{betaA}{Outcome model coefficient for I(A_i = 1) and I(A_i = 2)}

\item{betaX}{Outcome model coefficient vector for covariates X_i}

\item{betaXA}{Outcome model interaction coefficients for covariates. betaXA = c(betaXA1, betaXA2)}
}
\value{
a complete simulated data_frame
}
\description{
.. content for details ..
}
\author{
Kazuki Yoshida
}
