% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_generate_data.R
\name{generate_tri_treatment_from_two_covariates}
\alias{generate_tri_treatment_from_two_covariates}
\title{Generate three-valued treatment with constraint on}
\usage{
generate_tri_treatment_from_two_covariates(X, alpha01, alpha02, alphaXm1,
  alphaXm2, gamma, sigma)
}
\arguments{
\item{X}{data_frame containing two covariates}

\item{alpha01}{True intercept for the first linear predictor.}

\item{alpha02}{True intercept for the second linear predictor.}

\item{alphaXm1}{True coefficients for the first variable for the first linear predictor.}

\item{alphaXm2}{True coefficients for the first variable for the second linear predictor.}

\item{gamma}{Ratio of the true coefficient for the second variable and the first variable.}

\item{sigma}{scaling of all covariate effects. A higher value means stronger covariate effects, i.e., less clinical equipoise.}
}
\value{
df with treatment (A) added.
}
\description{
.. content for details ..
}
